Tidied camera class
- Tidied camera stuff
- Fixed delta time update
- General code tidy
- Reconstructed terrain constructor class, and fixed reloading shaders
with terrain

+++ CGR/CGR/src/anim_types.h
@@ -4,7 +4,7 @@
 #include "types.h"
 #include <vector>
 
-typedef float float3[3];
+typedef double double3[3];
 
 // Num pre-calculated normals
 #define NUM_VERT_NORMALS		162
